+++ src/main/resources/db/migration/V2__create_matches.sql
@@ -1,8 +1,9 @@
-CREATE TABLE matches (
-    id BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
-    match_id VARCHAR(32) NOT NULL UNIQUE,
-    patch VARCHAR(16) NOT NULL,
-    queue_id INT NOT NULL,
-    game_start TIMESTAMPTZ NOT NULL,
-    game_duration INT NOT NULL
+CREATE TABLE matches
+(
+    id            BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+    match_id      VARCHAR(32) NOT NULL UNIQUE,
+    patch         VARCHAR(16) NOT NULL,
+    queue_id      INT         NOT NULL,
+    game_start    TIMESTAMPTZ NOT NULL,
+    game_duration BIGINT      NOT NULL
 );
